Improve authenticated method.
Initialising an AuthenticationToken and providing the `profile_id` and `profile_name` parameters caused the authenticated property to return True, even though the Token has not been validated.

https://wiki.vg/Authentication#Validate
Validate
Checks if an `accessToken` is usable for authentication with a Minecraft server. The Minecraft Launcher (as of version 1.6.13) calls this endpoint on startup to verify that its saved token is still usable, and calls /refresh if this returns an error.